Move generic code to its own directory
