+++ opencog/persist/CMakeLists.txt
@@ -6,6 +6,10 @@ ADD_SUBDIRECTORY (file)
 #    ADD_SUBDIRECTORY (hypertable)
 # ENDIF (HYPERTABLE_FOUND AND HAVE_EXPAT)
 
+IF (GUILE_FOUND)
+	ADD_SUBDIRECTORY (guile)
+ENDIF (GUILE_FOUND)
+
 IF (ODBC_FOUND)
 	ADD_SUBDIRECTORY (sql)
 ENDIF (ODBC_FOUND)
