+++ opencog/persist/guile/CMakeLists.txt
@@ -1,0 +1,13 @@
+
+ADD_LIBRARY (persist SHARED
+	PersistSCM.cc
+)
+
+TARGET_LINK_LIBRARIES(persist
+	atomspace
+	smob
+)
+
+INSTALL (TARGETS persist
+	LIBRARY DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
+)
