+++ opencog/persist/guile/README.md
@@ -1,0 +1,5 @@
+                          Persist Guile API
+                          ------------------
+
+Guile wrappers for pesistence backends. Currently, the only backend is
+the SQL backend.
