+++ opencog/persist/sql/CMakeLists.txt
@@ -6,26 +6,26 @@ INCLUDE_DIRECTORIES (
 	${ODBC_INCLUDE_DIRS}
 )
 
-ADD_LIBRARY (persist SHARED
+ADD_LIBRARY (persist-sql SHARED
 	odbcxx.cc
 	AtomStorage.cc
-	PersistSCM.cc
 	PersistModule.cc
+	SQLPersistSCM.cc
 )
 
 # XXX FIXME We've added a depedency on the NLP atom types.
 # this should be removed when the NLP_HACK in PersistModule is removed.
-ADD_DEPENDENCIES(persist opencog_atom_types)
-ADD_DEPENDENCIES(persist nlp_atom_types)
+ADD_DEPENDENCIES(persist-sql opencog_atom_types)
+ADD_DEPENDENCIES(persist-sql nlp_atom_types)
 
-TARGET_LINK_LIBRARIES(persist
+TARGET_LINK_LIBRARIES(persist-sql
 	atomspace
 	nlp-types
 	${ODBC_LIBRARIES}
 )
 
 IF (HAVE_GUILE)
-	TARGET_LINK_LIBRARIES(persist smob)
+	TARGET_LINK_LIBRARIES(persist-sql smob)
 ENDIF (HAVE_GUILE)
 
 ADD_EXECUTABLE(sniff
@@ -33,19 +33,12 @@ ADD_EXECUTABLE(sniff
 )
 
 TARGET_LINK_LIBRARIES(sniff
-	persist
+	persist-sql
 	server
 	atomspace
 	${ODBC_LIBRARIES}
 )
 
-INSTALL (TARGETS persist
+INSTALL (TARGETS persist-sql
 	LIBRARY DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
-)
-
-INSTALL (FILES
-	AtomStorage.h
-	odbcxx.h
-	PersistModule.h
-	DESTINATION "include/${PROJECT_NAME}/persist/sql"
 )
