+++ opencog/server/CMakeLists.txt
@@ -49,7 +49,7 @@ IF (HAVE_GUILE)
 ENDIF (HAVE_GUILE)
 
 IF (ODBC_FOUND)
-	TARGET_LINK_LIBRARIES(server persist)
+	TARGET_LINK_LIBRARIES(server persist-sql)
 ENDIF (ODBC_FOUND)
